Add gnu-arm-embedded-toolchain to CCI
Specify library name and version:

    - gnu-arm-embedded-toolchain/11.3.0
    - gnu-arm-embedded-toolchain/12.2.0

New addition to ConanCenter.

The author of the toolchain binaries is ARM inc and the archive files
can be found here:
https://developer.arm.com/downloads/-/arm-gnu-toolchain-downloads

This recipe uses a github repo releases rather than the links from the
arm website as their security/DDOS measures make it harder to work with
in conan, and download speeds are much slower compared to github.

This package exists to make installing the toolchain files easier when
using conan and to make the packages easily available via the
[tool_requires] section of a project.