(#15254) openimageio: add version 2.4.7.1
* first minimal patch

* don't use cmake modules, fix tsl-robinmap

* only call checked_find_package for optional dependencies if option is activated

* fix: add 2.4.7.1 to config.yml

* fix: add missing endline

* fix: use more recent package versions, deactivate OpenCV by default

OpenImageIO depends on OpenColorIO and OpenCV, which both depend on OpenEXR. OpenColorIO in Conan Center are build using openexr/2.5.7 and OpenCV packages are build using openexr/3.1.5. Because of that, OpenImageIO won't build in CI, because of missing packages

* fix: address linter issues

* fix: linker error in test_package

* disable old versions to see if CI gets through

* fix: ffmpeg and ptex where not found

* disable old versions to see if CI gets through

* fix yaml linter issues

* try to reenable OpenCV by default

* disable OpenCV by default

OpenImageIO depends on OpenColorIO and OpenCV, which both depend on OpenEXR. OpenColorIO in Conan Center are build using openexr/2.5.7 and OpenCV packages are build using openexr/3.1.5. Because of that, OpenImageIO won't build in CI, because of missing packages

* temporarily disable tbb, to see if CI is passing for everything else

* try: workaround linker error on windows

* fix: find ffmpeg

* fix: find ptex

* reenable older versions

* add missing information about the patch

* split MSVC patch into separate file

* remove comment

* refractor: use apply_conandata_patches helper to apply patches