apr: Loosen the cross-building restriction
From all the discussion I found in #6622, the issue which motivated a
"doesn't support cross-build yet" is the use of is AC_TRY_RUN checks
in ./configure. That's not an issue when consuming existing binaries,
so this check belongs in validate_build (since conan 1.51.0).

It also doesn't apply to MSVC which uses CMake rather than autoconf;
There's no use of try_compile or try_run in the CMake project.