(#16491) (#16490)[package] nas: Fix recipe on ARM platforms
* (#16490)[package] nas: Fix recipe on ARM platforms

The repository contained an old version of the config.guess and config.sub files, which are not compatible with building on ARM. During the configure phase it is unable to guess the system type. Overwriting both files with a new version fixes that and makes it recognize the system correctly.

* Update recipes/nas/all/conanfile.py

Co-authored-by: SpaceIm <30052553+SpaceIm@users.noreply.github.com>

* Update recipes/nas/all/conanfile.py

Co-authored-by: SpaceIm <30052553+SpaceIm@users.noreply.github.com>

---------

Co-authored-by: SpaceIm <30052553+SpaceIm@users.noreply.github.com>