(#17197) libffi/3.4.3: Add patch to allow building with the C2x standard
This patch allows building with newer versions of Clang out-of-the-box.
Clang 15.0.0 removed support for implicit function declarations.
This is mentioned here: https://releases.llvm.org/15.0.0/tools/clang/docs/ReleaseNotes.html#c2x-feature-support.
This removal breaks builds using the default C standard in Clang.
This is because the function open_temp_exec_file is not properly declared.

This patch is from the upstream project and declares the function.