libuuid/1.0.3: Fix missing include for the sys/file.h header
This fixes compilation when using Clang 16.
Due to the removal of support for implicit declarations, libuuid fails to build due to an issue where the `<sys/file.h>` header is not included.
Specifically, the `flock` function requires `<sys/file.h> to be included.
Since the compiler preprocessor macro `HAVE_SYS_FILE_H` is not defined, that header is not included.
This worked previously since the compiler assumed that is was an implicit function declaration.
Defining the macro on the necessary platforms solves the issue.