(#16782) apr: allow recipes which support cross_building to consume exising apr binaries
* apr: Remove settings_build workaround for tools.cross_building

As discussed in https://github.com/conan-io/conan-center-index/pull/6622#discussion_r685814490
conan 1.38.0 and up have the new conan.tools.build.cross_building which
compares build vs host settings, without the fallback on os_build/arch_build
that the old tools.cross_compiling has.

The recipe was converted in #14797 to use conan.tools.build.cross_building
and required_conan_version = ">=1.53.0", making the workaround is obsolete.

* apr: Loosen the cross-building restriction

From all the discussion I found in #6622, the issue which motivated a
"doesn't support cross-build yet" is the use of is AC_TRY_RUN checks
in ./configure. That's not an issue when consuming existing binaries,
so this check belongs in validate_build (since conan 1.51.0).

It also doesn't apply to MSVC which uses CMake rather than autoconf;
There's no use of try_compile or try_run in the CMake project.

---------

Co-authored-by: Rubén Rincón Blanco <rubenrb@jfrog.com>