(#17355) avahi/0.8: Disable the use of the setproctitle function on Linux
* avahi/0.8: Disable the use of the setproctitle function on Linux

This function is not declared on Linux.
Configure ignores warnings about the missing declaration.
This causes issues when building with recent versions of Clang.
Support for implicit declarations was removed in newer C standards in Clang.
Telling configure that the setproctitle function is not available disables the use of this function in Avahi.

* Capitalize Linux