(#17691) arrow: fix the library name for the libacero component
I think we got this part wrong in https://github.com/conan-io/conan-center-index/pull/17540
It's not clear if this name stays (see https://github.com/apache/arrow/issues/15280) but currently
the file being created is libarrow_acero.a

I'm not certain what to do with names of other components, such as pkg_config: for arrow_substrait
we prepend "arrow_" everywhere, for "arrow_flight_sql" we do not. I've chosen to make
the smallest possible diff that I am sure about.

Co-authored-by: James <james@conan.io>