Ruby: support conan v2 and expose more configuration options for native extension gems. Reboots #12456 and #12208
* conan v2 support (work done by myself + a merge of @SpaceIm 's branch from PR #12208 to reconcile both)

* Extend testing to require one of the native extensions

* ```
post_package(): WARN: [MISSING SYSTEM LIBS (KB-H043)] Library './lib/ruby/3.1.0/x86_64-linux/pty.so' links to system library 'util' but it is not in cpp_info.system_libs.
```

* Use `--with-opt-dir` instead of `--with-xx-dir` that ruby isn't respecting.

conanio/gccXX (eg 10) removed the libxxx-dev (eg libgmp-dev) from the image. This made me realize that the conan deps weren't being picked up by the build. The openssl was, because --with-openssl-dir is explicitly used in ruby config.

So here, we rely on --with-opt-dir.

cf: https://bugs.ruby-lang.org/issues/19014#change-99241 (--with-gmp-dir was added on September 14, 2022, wrote this in Sep 22, 2022)

* Pass both --with-opt-dir and --with-xxx-dir

https://gist.github.com/mrkn/6647630