(#19263) libgettext/*: Set license as LGPL instead of GPL
* libgettext/*: Set license as LGPL instead of GPL

The portions of libgettext that we package are licensed under the LGPL, not the GPL.
We only package the libgnuintl portion of the libgettext package, which is under the LGPL.
This can easily be confirmed by checking the packaged libintl.h header file.
This is also documented explicitly by the project here: https://www.gnu.org/software/gettext/manual/gettext.html#Licenses.

* Add comment about the licensing

---------

Co-authored-by: Carlos Zoido <mrgalleta@gmail.com>