[armadillo] Prevent static builds when use_wrapper=True
For armadillo to wrap other libraries, these other libraries must be
either statically compiled into the armadillo binary or dynamically
linked to by libarmadillo. This is not possible when armadillo is a
static library itself, without doing some unusual modification of the
armadillo archive.

This change ensures that armadillo is built as a shared library when
use_wrapper=True, which is the only configuration that would enable
libarmadillo to be used this way.

Closes #7840