[armadillo] Propagate shared obejct requirements transitively
This is necessary to avoid linker errors as of binutils 2.22, which gives
errors similar to `libhdf5.so.310: error adding symbols: DSO missing
from command line`. This more strict requirement in binutils requires
that `-lhdf5` or `-lopenblas` be added to the consumer linker command
line invocation, which passing these libraries transitively does.