yaml linter: remove warning about base_path
base_path is necessary in some cases (eg qt) where upstream patches apply to only a sub-directory of the sources. This happens when the sources are an aggregation of several repositories.

the presence of base_path is an indirect consequence of explicitly setting the `destination` argument  of the `get` function,  so a better diagnostic would be to detect these.