(#19907) [armadillo] Add version 12.6.4
* [armadillo] Add version 12.6.4

* [armadillo] Propagate hdf5 headers transitively

This is necessary because armadillo_buts/include_hdf5.h includes the hdf5 headers.
When a consumer links armadillo with hdf5 and then includes the
armadillo headers, they need to be able to discover them.

* [armadillo] Propagate shared obejct requirements transitively

This is necessary to avoid linker errors as of binutils 2.22, which gives
errors similar to `libhdf5.so.310: error adding symbols: DSO missing
from command line`. This more strict requirement in binutils requires
that `-lhdf5` or `-lopenblas` be added to the consumer linker command
line invocation, which passing these libraries transitively does.