(#19909) [armadillo] Prevent static builds when use_wrapper=True
* [armadillo] Prevent static builds when use_wrapper=True

For armadillo to wrap other libraries, these other libraries must be
either statically compiled into the armadillo binary or dynamically
linked to by libarmadillo. This is not possible when armadillo is a
static library itself, without doing some unusual modification of the
armadillo archive.

This change ensures that armadillo is built as a shared library when
use_wrapper=True, which is the only configuration that would enable
libarmadillo to be used this way.

Closes #7840

* [armadillo] Prevent transitive library propagate when use_wrapper=True

The armadillo run-time library wrapper is designed to provide an opaque
integration with upstream libraries, and so these are not required to be
passed transitively.

* Bump hdf5 version

Co-authored-by: SpaceIm <30052553+SpaceIm@users.noreply.github.com>

---------

Co-authored-by: SpaceIm <30052553+SpaceIm@users.noreply.github.com>