(#19593) libdwarf: add version 0.7.0, fix missing header in v0.5.0
* libdwarf: Fix missing dwarf.h in package v0.5.0

Updated CMake patch to also install dwarf.h. The patch was already in
place for version v20191104 but was missed when the recipe got updated
to v0.5.0.

Also updated tests to use Dwarf Attributes defined in dwarf.h.

* Remove dependency on libelf if build_dwargen=False

libdwarf doesn't depend on libelf since version 0.1.0. The dependency is
only needed if dwarfgen is built.

The patch drops the REQUIRED parameter for find_package on libelf, it
should be available only when build_dwarfgen=True.

Fixed syntax error in conanfile when calling append.

* libdwarf: add version 0.7.0

* Check BUILD_DWARGEN before looking for libelf

Instead of dropping the REQUIRED options from find_package, don't search
for it if we don't need it.

* Stop exporting libdwarf.pc file

---------

Co-authored-by: Uilian Ries <uilianries@gmail.com>