(#20721) [google-cloud-cpp] bump requirements and other fixes
* [google-cloud-cpp] bump requirements and other fixes

Several requirements need version bumps. Where possible I chose
version ranges over pinning specific versions.

I had to backport fixes for MSVC. I am only guessing, but it seems
the toolchain changed in the CI system and now spelling `||` as `or`
results in build errors.

I also disabled `-Werror` because the code in `v2.5.0` has warnings
with newer toolchains. This is to be expected, and that is why the
project no longer enables `-Werror` by default.

* Require host version for gRPC

* Starting with v2.12.0 we need a different fix on macOS

* Address review comments