[openblas] Make build_lapack=True default for >= 0.3.21
Version 0.3.21 of openblas introduced the ability to build lapack from
C. Make this default to enable CCI recipes to use lapack when depending
on openblas.

Additionally, fix an inconsistency in version constraint. Review was
provided in https://github.com/conan-io/conan-center-index/pull/19808 to
constrain it to 0.3.24 as this was the latest version available in CCI.
However, the constraint on building lapack really applies to anything
newer than 0.3.21, so this is the most appropriate model for the logic.
Not modelling this accurately precludes the correct behaviour if
intermediate versions are added at a later date.

Closes #7361