#21292 dat_package_file option improved.
dat_package_file option usage was impractical for consumer recipes. It required that the file specified for the dat_package_file option to exist on the consumer machine since the hash of the file is calculated in package_id method and existence of the file was verified by the validate method.

Introduced a new custom_data_hash option which is calculated from the hash of the file specified in dat_package_file option.
package_id method calculates the hash of the file and sets the custom_data_hash option accordingly and removes the dat_package_file option.

During package creation the dat_package_file option shall be set to the appropriate data file.
In consumer recipes the dat_package_file option cannot be used.

In this way the consumer recipe does not require the file specified in the dat_package_file option to exist but icu built from custom data file can still be used.