(#22081) xorg: remove xvmc as dependency
xvmc is the X-Video Motion Compensation API client library. The xvmc
library is not referenced in any other package in
conan-center-index. The libXvMC-devel has unceremoniously been removed
from RHEL 9. A github search for xvmc finds 15 repositories and not
one of them seem to be an important consuming application of this
API. libxcb-devel contains libxcb-xvmc.so which provides an XCB based
API.

As the xvmc library is unused in conan-center-index, appears to be
nearly unreferenced in github and has been removed from a major
distribution it should not be included in xorg/system either so
xorg/system can be built on RHEL 9.

Documentation for the removal of libXvMC in RHEL 9:

https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html-single/considerations_in_adopting_rhel_9/index