nodejs: fix glibc version check on Fedora, RHEL and probably more
The glibc version check implemented in commit 89fa4810f031 wont work
on many distributions. The implemented check finds the version number
by scanning the output of `ldd --version` with the regex /GLIBC
(\d{1,3}.\d{1,3})/ - this works for debian and ubuntu, but not for
Fedora, and RHEL/CentOS:

    $ podman run debian:latest ldd --version |head -1
    ldd (Debian GLIBC 2.36-9+deb12u3) 2.36

    $ podman run ubuntu:latest ldd --version |head -1
    ldd (Ubuntu GLIBC 2.35-0ubuntu3.5) 2.35

    $ podman run centos:7 ldd --version |head -1
    ldd (GNU libc) 2.17

    $ podman run fedora:latest ldd --version |head -1
    ldd (GNU libc) 2.38

Note that the output format "ldd (GNU libc) <version>" is the default
for the ldd tool from binutils. Replacing "GNU libc" with a different
string is done by defining PKGVERSION when building binutils which
means the presence of GLIBC in this text is entirely dependent on the
Linux distribution.

The problem could be fixed by parsing the version number at the end of
the string, but I think it is better to use Python's
platform.libc_ver() function.