(#20894) [openblas] Make build_lapack=True default for >= 0.3.21
* [openblas] Make build_lapack=True default for >= 0.3.21

Version 0.3.21 of openblas introduced the ability to build lapack from
C. Make this default to enable CCI recipes to use lapack when depending
on openblas.

Additionally, fix an inconsistency in version constraint. Review was
provided in https://github.com/conan-io/conan-center-index/pull/19808 to
constrain it to 0.3.24 as this was the latest version available in CCI.
However, the constraint on building lapack really applies to anything
newer than 0.3.21, so this is the most appropriate model for the logic.
Not modelling this accurately precludes the correct behaviour if
intermediate versions are added at a later date.

Closes #7361

* [openblas] Switch option change to config_options from configure

Modifying defaults in configure is now prohibited and should be done in
config_options instead.

* [openblas] Only propagate -lgfortran when compiled with fortran

* Add note to explain lapack source when no Fortran compiler is available

Co-authored-by: Uilian Ries <uilianries@gmail.com>

* [openblas] set TARGET for armv8 builds to facilitate architecture detection

* [openblas] Add armv8 compatibility

* Patch versions earlier than 0.3.24 to include hotfix of
  OpenMathLib/OpenBLAS#4142, which fixes detection of the armv8
  architecture to ARM64 from ARM.
* Refactor patches into a _patch_sources method

* Revert 0bd65ec and b42908

* [openblas] Disable armv8 builds for versions lower than 0.3.24.

* armv8 is not currently supported by versions lower than 0.3.24 due to
  a bug in the cpu architecture identification erroneously identifying a
  32 bit architecture instead of 64.

* [openblas] Re-add erroneously removed version check for when to build lapack

This check was erroneously removed when the note about an f2c converted
copy of lapack being used was added. The f2c version of lapack is only
available for versions greater than or equal to 0.3.21, and versions
lower than this should not build lapack by default.

---------

Co-authored-by: Uilian Ries <uilianries@gmail.com>