[armadillo] Add alias targets for wide compatibility
Add alias targets for wide compatibility with existing CMake scripts.
The upstream CMake find module doesn't define what the name of the
targets should be, so this provides coverage for the wide usage of
different possible combinations found in the wild.

This is based on inspection of target_link_library patterns for
armadillo usage found on github.