[tz] Refactor self.dependencies usage for conan 1.x compatibility
Conan 1.x is not able to parse self.dependencies in the test() method
due to differences in the lifetime of objects. The solution presented
here is to save the necessary options to file and load them back again
in the test() method to be used as appropriate.

Additionally, utilise the conan `save` and `load` methods for
readability and consistency with other recipes.

Co-authored-by: Martin Valgur <martin.valgur@gmail.com>