avahi: Fix conflicting definition of strlcpy from glibc
I was unable to coerce glibc to not include the function declaration for strlcpy.
It's included in `/usr/include/string.h` in recent versions of glibc.
At least version 2.38 of glibc.
Since I couldn't avoid the definition from glibc, I removed the static keyword to make the definitions the same.
Avahi should still use its internal strlcpy in this situation.