avahi: Support cross-compilation
To cross-compile Avahi, the --with-distro flag must be passed.
Disable introspection.
This must be disabled to cross-compile.
Strangely enough, it doesn't work when enabled.
It only seems to work when set to auto.
When gobject-introspection is updated to support Conan V2 an option could be added to control enabling or disabling introspection.