libva/2.20.0: Fix discovery of wayland-scanner
Cross-compilation is currently broken for the libva package for Conan V1.
It doesn't properly make the wayland package available in the build context through pkg-config.
This is necessary for the project to find the wayland-scanner program.
I've updated libva to use the same procedure as xkbcommon to incorporate Wayland's pkg-config in the build context.

Note that this change is not necessary for Conan version 2.2.1 which works out-of-the-box.