llvm-core: unpin cmake version
In llvm-core, the cmake version was pinned to 3.20.5, making it
impossible to build with VS 2022 on Windows since VS 2022 generators
were introduced with cmake 3.21. From the original comment, it seems the
intention was not to pin cmake, but rather specify a minimum version.

This change simply transforms the cmake pinning to a mere minimum
requirement, making it possible to compile llvm-core with toolchains
born after cmake 3.20.5.