libepoxy: Use libglvnd instead of opengl/system and egl/system
The libglvnd package provides the necessary opengl/system and egl/system dependencies.
The libglvnd package is a proper, non-system Conan package.
Bump Meson to 1.4.0.
Require the x11 option to be enabled when the glx option is enabled.
This is required in the meson.build file of the project.
Support FreeBSD.