(#23010) fontconfig: Remove patch for freetype libtool version number
* fontconfig: Remove patch for freetype libtool version number

This patch is a bad workaround for improper versioning in the freetype package.
This is fixed in the freetype package itself in #22974.

* Fix missing conf fontconfig file

This fixes the following error, which can be seen when running the test package:

```
Fontconfig error: Cannot load default config file: No such file: (null)
```

This is fixed by installing the fontconfig configuration file.
It was being installed to the wrong location in the Meson package previously.
Additionally, don't set `FONTCONFIG_FILE`, but instead append to `FONTCONFIG_PATH` only.
This works and makes it easier for consumers to opt for the system font directory.
See #5782.

* Use relative paths for the updated Meson generator