(#22952) libselinux: Fix cross-compilation
* libselinux: Fix cross-compilation

Apply patches for 3.6 to fix detection of available functions when cross-compiling.
Use the PKG_CONFIG_LIBDIR to point pkg-config to the correct directory containing Conan's generated pkg-config files.
Without this, the pcre2 from the system will be picked up before the one provided by Conan.
When cross-compiling, this causes system include directories to poison the build.

* Bump pkgconf

* Only set PKG_CONFIG_LIBDIR when cross-compiling

* Update recipes/libselinux/all/conanfile.py

---------

Co-authored-by: Francisco Ramírez <franchuti688@gmail.com>