Fixing error specific to versions < 1.3.3, and the none check
In older versions of tomlplusplus there was no .table function on result_type when exceptions were disabled. I've fixed that by using a cast to table& wwhich is supported on older versions.

I also fixed an issue with the None check on the exceptions option which is now check equality rather than identity. I've tested that this is working with Conan 2.4.1.