(#24336) Exposing the exceptions feature of tomlplusplus
* Exposing the exceptions feature of tomlplusplus

* Fixing linter error C0304 Final newline missing

recipes\tomlplusplus\all\conanfile.py:84:0: C0304: Final newline missing (missing-final-newline)

* Fixing error specific to versions < 1.3.3, and the none check

In older versions of tomlplusplus there was no .table function on result_type when exceptions were disabled. I've fixed that by using a cast to table& wwhich is supported on older versions.

I also fixed an issue with the None check on the exceptions option which is now check equality rather than identity. I've tested that this is working with Conan 2.4.1.

* linter did not like "!=" over is not none.

I've fixed the "is not None" to check the value which is now correctly seeing None when the default value is set.