cmake: re-add from source builds as `cci.` versions
I think that the decisions to re-package binary downloads was a mistake to begin with. Most of the named arguments are also outdated today as they are solved by version ranges.

This PR makes the CMake recipe that builds from source again available from the Conan Center remote under the `cci.` version prefix.