Update conandata_yml_format.md
When reading the documentation here I misread that it was the hash of the file rather than the url, looking back this is an obvious mistake because we care about the files matching not the URL's matching. 

I decided to make it even more obvious for anyone else who might have gotten confused in the moment.