Fix shared library of zlib
When option.shared = True and compiler is MSVC, the CMakeLists.txt
creates an archive called zlib.lib, but the dll is called zdll.dll.

This configuration is not standard and is not supported by
cmakedeps_macro.cmake module when using find_package(ZLIB).

The source patch now creates zlib.lib and zlib.dll.