(#25517) Fix: incorrect headers in prebuilt package
Due to how `package_id` is implemented, when `_is_header_only()` returns True, the package_id will be the same for both values of `with_rapidyaml`. In addition to that, it seems conan-center contains prebuild version of the package_id that was created with `with_rapidyaml=False`. In effect, it contains broken headers for builds with `with_rapidyaml=True`

This fix makes sure we always patch header related to rapidyaml (it is harmless if this functionality is disabled).

NOTE TO CONAN-CENTER MAINTAINERS: To properly complete the fix, the current prebuild packages in conan-center must be replaced with fresh versions build with updated conanfile.py.