Stop supporting 1.4.4 and improve Windows support.
The patch version 1.4.4 has several limitations and works diffrently
from newers. As it's a patch version, stopping its support in Conan
Center will simplify the recipe.

When building on Windows, the AUTOMOC option is enabled by default, and
application Qt moc is not found in case the PATH is not well configured
the VirtualRunEnv is the answer. The same reflects by adding Qt as
tool_requires too, otherwise, there will a problem when cross-building
the project.

When building as shared libray on Windows, the define QXlsx_SHAREDLIB is
marked as PUBLIC in the upstream's CMakelists.

The upstream added QXlsx as target alias too.

Signed-off-by: Uilian Ries <uilianries@gmail.com>