Add version 0.4.0, 0.4.1 of Au to conan
Unfortunately, we discovered a small breakage in 0.4.0: we forgot to add
one of the headers to `CMakeLists.txt`.  I added a patch which fixes the
build for 0.4.0.  Version 0.4.1 does not use this patch.

Along the way, I also used Au's new options to avoid enabling tests and
bringing in the GTest dependency.  This means we don't need to keep
using a version of the existing patch file for 0.3.5.

I also updated the test file to be more idiomatic.

I tested these installations with the following commands:

```sh
cd recipes/au
conan create all/conanfile.py --version=0.4.0
conan create all/conanfile.py --version=0.4.1
```

Fixes #26181.