Fixed an issue where the wrong ZLIB library could be specified during OpenSSL configuration.
In #18011, it was demonstrated that the official Conan recipe for
OpenSSL could result in a build misconfiguration, leading to a "Bad
Image" message box appearing at runtime. This was ultimately the
result of attempting to load a static library as a DLL. I confirmed
this by placing a breakpoint on the following Win32 APIs:

  * `LoadLibraryA()`
  * `LoadLibraryW()`

These are the standard Win32 APIs for injecting a DLL into the
address space of the current process at runtime. (For more
information, refer to the MSDN at
https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibrarya.)
Checking the call stack in a debugger just before the message box
would appear revealed that an attempt was being made from within a
module by the name of `libcrypto-3-x64.dll` to load a static
library (`zdll.lib`) as a DLL. Such an operation is inherently
malformed, which is why the "Bad Image" message box would appear.

As it turns out, the issue was caused by an error in the
`conanfile.py` for OpenSSL. When ZLIB was built as a shared
library, the `OpenSSLConan::_configure_args()` function would
historically append the following command line arguments to the
list used to configure OpenSSL:

  * `zlib-dynamic`
  * `--with-zlib-lib=[Path to Static Import Library]`

Here, `[Path to Static Import Library]` represents the path to the
static import library which is generated alongside ZLIB's DLL in a
shared build of the project. When using the official Conan recipe
for ZLIB, this file is called `zdll.lib`, which - not
coincidentally - is the same library as the one which
`libcrypto-3-x64.dll` was erroneously attempting to load as a DLL.
Still, it is worth asking why this behavior occurred in the first
place. After all, even when a shared library is built, the linker
may still require access to the corresponding static import library
`zdll.lib`.

Well, the issue appears to have been caused by a misunderstanding
of the `--with-zlib-lib` command line argument. To be more precise,
it is important to understand how the `zlib-dynamic` command line
argument affects how it is interpreted. Inspecting the `Configure`
file located at the root of the OpenSSL source code repository
reveals the following crucial information contained in a comment:

```
[no-]zlib     [don't] compile support for zlib compression.
zlib-dynamic  Like "zlib", but the zlib library is expected to be a shared
              library and will be loaded in run-time by the OpenSSL library.
```

The implication of the above is that the `--with-zlib-lib` command
line argument is expected to be used to specify the ZLIB DLL
(typically `zlib1.dll`) when the `zlib-dynamic` command line
argument is also specified, rather than the corresponding static
import library (`zdll.lib`).

This commit addresses that issue by modifying the
`OpenSSLConan::_configure_args()` function so that the correct
library is specified on Windows for the `--with-zlib-lib` command
line argument. As a result, `libcrypto-3-x64.dll` should now
attempt to load the correct DLL for ZLIB at runtime. At a minimum,
this prevents the aforementioned "Bad Image" message box from
appearing. One might also speculate what other subtle OpenSSL bugs
may have been fixed by this change.

NOTE: The changes in this commit do _not_ modify the library
specified via the `--with-zlib-lib` command line argument on Linux.
A comment within `OpenSSLConan::_configure_args()` suggests that
the "right file" will be specified, but it might be worth verifying
this claim in response to the findings described above.

This commit fixes #18011.