Add xtl 0.8.0 version
This version will be required for bumping the xtensor version to 0.26.0.

The version 0.8.0 required some changes to the conanfile.py. Starting
from this version a cppstd >= 17 is required. This also means, that
newer compiler versions will be required. I took those exact versions
from the cppreference and looked at other packages that use cppstd >=
17.

Further, the xvariant header was removed from that version. So the
test_package was updated to test with xoptional instead, which is
available in every version.