Add xtensor 0.26.0 version
This version adds support for clang 19 and above.

The version 0.26.0 required some changes to the conanfile.py. Starting
from this version a cppstd >= 17 is required. This also means, that
newer compiler versions will be required. I took those exact versions
from the cppreference and looked at other packages that use cppstd >=
17.

Further, code was restructured between this and the last version and
with that include paths changed. The changes to the test_package make
sure that the test still works with all versions.