Build shared library again
The upstream confirmed that's possible to build as shared library and
there is no block, but it was a decision that can be reverted. The
project itself has no limitation in terms of building, so I
re-introduced the shared option. Of course, on Windows it will not work
properly because it's not exporting symbols.

Signed-off-by: Uilian Ries <uilianries@gmail.com>