recipes/ffmpeg/all/conanfile.py: support drawtext avfilter
freetype and harbuzz (6.1+) are required options for the drawtext
avfilter.

fontconfig is optional and allows font fallback and
font configuration. fribidi is optional but can improve
rendering.

https://ffmpeg.org/ffmpeg-filters.html#drawtext-1

all are disabled by default to preserve the previous behavior
of the conan ffmpeg package.

harfbuzz version is set to 8.3.0 to reduce conflicts since most
of Conan center is pinned to 8.x.

https://github.com/conan-io/conan-center-index/issues/22578