botan: don't use the amalgamation by default (#28010)
The amalgamation build is useful to integrate Botan as a single header/cpp file
into downstream projects. When using the library through Conan it does not pose
a relevant advantage. On the contrary, we saw miscompilations when using GCC 13
with the amalgamation. Hence, it does not seem to be a good default.

See also: https://github.com/randombit/botan/pull/5011