Updated the `osgearth` recipe to support v3.7.2.
The official Conan recipe for `osgearth` (located at
`recipes\osgearth\all\conanfile.py`) appears to have been neglected
for a long time. Consider, for example, the following:

  * There existed several dependency conflicts based on the
    versioned references specified in the `osgearth` recipe.

  * OpenSceneGraph and GDAL were not building on the latest version
    of the MSVC. (These are both mandatory dependencies of
    `osgearth`.)

  * Several of the member functions of `OsgearthConan` made use of
    old Conan 1 facilities, such as the old `conans.CMake` type,
    which are either deprecated or removed in Conan 2.

This commit updates the `OsgearthConan` class (that is, the
official Conan recipe for `osgearth`) to support Conan 2 in
addition to a new version of the library - v3.7.2. (Support for
both v3.2 and v3.3 has been retained in this commit. However, this
is subject to change pending a code review.) This required a
significant number of changes to be made to its implementation.
Along the way, several bugs in the previous implementation were
encountered and fixed. These are as follows:

  * The `OsgearthConan::generate()` member function will now set
    the `CMAKE_*_POSTFIX` CMake variables in the CMake cache in
    order to match what is expected by osgDB. Specifically, any
    plugins which are loaded by osgDB in a shared/DLL build of
    OpenSceneGraph are expected to have a postfix of
    `OSG_LIBRARY_POSTFIX` in their file name. This is a
    preprocessor macro which is set to match the value of
    `CMAKE_*_POSTFIX` during the build of osgDB; since
    OpenSceneGraph uses a particular set of values for the
    `CMAKE_*_POSTFIX` variables, we want the osgDB plugins
    generated by `osgearth` to also use these values.

  * For some reason, the `OsgearthConan::requirements()` member
    function would add `zstd` as a dependency when the
    `build_zip_plugin` option was set to `True`. In reality,
    however, the `osgdb_zip` CMake target relies on `libzip`, and
    not `zstd`. This dependency has thus been adjusted accordingly
    in this commit.

  * The official Conan recipe for `rocksdb` will delete the C++
    headers from the packaged include directory. (A comment claims
    that this is done for the sake of ABI stability.) However,
    `osgearth` relies on these headers to exist in order to build
    the `osgdb_osgearth_cache_rocksdb` CMake target. This means
    that we cannot use a shared build of `rocksdb` in order to
    build that particular plugin. Thus, this commit updates the
    `OsgearthConan::validate()` member function to check for this
    particular case.

There are some additional caveats which I would like to mention
here. `osgearth` includes an optional `imgui` integration/nodekit
which has historically been very problematic. Consider, for
example, the confusing evolution of how this integration was
distributed:

  * `osgearth` v3.2 includes an embedded version of `imgui` (and
    `imnodes`) which is used to build the `imgui` nodekit, but only
    if GLEW exists as a dependency.

  * `osgearth` v3.3 removes the embedded version of `imgui` (and
    `imnodes`), effectively rendering the `imgui` nodekit unusable
    (that is, unless the user actually copies their own version of
    `imgui` into the source directory).

  * `osgearth` v3.7.2 yet again includes an embedded version of
    `imgui` (and `imnodes`), and yet again, GLEW must exist as a
    dependency in order to build the `imgui` nodekit. This time,
    however, there is an additional requirement: the
    `OSGEARTH_BUILD_IMGUI_NODEKIT` CMake variable must also be set.

It should be possible to patch `osgearth` to ensure consistency
between all of our supported versions. However, this is expected to
take a significant amount of effort. This commit thus always
disables the `imgui` nodekit. Future work would be to add optional
support for it in the form of a proposed `build_imgui_nodekit`
option.

There also appears to be some issues with the
`build_procedural_nodekit` option in older versions of `osgearth`.
Specifically, the `osgEarthProcedural` CMake target does not seem
to build on the MSVC for these versions. The
`OsgearthConan::config_options()` member function seems to check
for this, but doing so there is arguably pointless because any
changes made to that option from within
`OsgearthConan::config_options()` will be overwritten by the value
specified in, e.g., the current Conan profile. No explanation is
given as to why this check is done, but it seems incorrect. (It
should be noted that this does not appear to be an issue for
`osgearth` v3.7.2; the `osgEarthProcedural` CMake target builds
fine on my machine with the MSVC in that version.)

Fixes: #25927