Updated the `test_package` recipe for `osgearth`.
This commit updates the testing recipe for `osgearth`. This recipe
is located at `recipes\osgearth\all\test_package\conanfile.py`.
The root `CMakeLists.txt` for this testing recipe has been updated
to resemble the examples given in
[the Conan 2 documentation](https://docs.conan.io/2/tutorial/creating_packages/test_conan_packages.html),
as has the `TestPackageConan` class defined in
`recipes\osgearth\all\test_package\conanfile.py`.

During this testing, it was revealed that the `test_package` CMake
target managed by the testing recipe failed to build. This was
ultimately because Conan does not make the headers of dependencies
transitive by default; this must be explicitly requested when
calling the `ConanFile::requires()` member function by setting the
`transitive_headers` parameter to `True`. Since the public headers
of `osgearth` will frequently attempt to `#include` the public
headers of `openscenegraph`, it makes sense to use transitive
headers for that particular dependency. This conclusion resulted in
the `OsgearthConan::requirements()` member function being updated
accordingly in this commit.