Removed support for older versions of `osgearth`.
This commit removes support for the following versions of
`osgearth` from its official Conan recipe:

  * v3.2
  * v3.3

This change came from a suggestion by @AbrilRBS that the older
versions could be removed if doing so would simplify the final
`osgearth` recipe. As can be seen from the changes in this commit,
this did, in fact, result in a much simpler implementation for the
`OsgearthConan` class. The `CMakeLists.txt` file which was
previously being injected into the `osgearth` build (via
`OsgearthConan.exports_source`) has been removed in this commit, as
the CMake scripts used in `osgearth` v3.7.2 make proper use of the
`find_package()` CMake command (in most cases, anyways).

This commit also deprecates the following options, as described in
[the FAQs documentation](https://github.com/conan-io/conan-center-index/blob/master/docs/faqs.md#can-i-remove-an-option-from-a-recipe):

  * `build_leveldb_cache`
  * `with_draco`

These options are ultimately useless in `osgearth` v3.7.2. In
previous versions, they were used to set the values of various
CMake variables, but said variables no longer exist.