Modified the shader installation directory for `osgearth`.
The official Conan recipe for `osgearth` would historically install
shaders into a directory called `res`, located at the root of the
package folder. During a code review, however, @AbrilRBS asked that
the shader installation directory used by `osgearth` be retained,
rather than renamed to match the behavior of previous versions of
the `osgearth` recipe. This is because the old recipe had not been
updated to Conan 2 prior to #28101.

Based on that feedback, this commit updates the
`OsgearthConan::package()` member function to no longer attempt to
rename the shader installation directory. For all intents and
purposes, future recipe versions should consider the `share` folder
as the destination for installed shaders.