Fixed several `openscenegraph` build failures.
This commit includes several additional patch files in order to
resolve build failures which occurred when attempting to build
`openscenegraph` v3.6.5 with the MSVC. An adjustment was also made
to the declared required `libjpeg` version; this was done to
resolve a dependency conflict for #28101.

I should mention that I have not done any significant analysis
regarding the implications of increasing the `libjpeg` version
requirement. This change resolves the aforementioned dependency
conflict, but it could also theoretically introduce new ones. It
may be better to replace this with a version range, instead.