Fixed some dependency conflicts with `osgearth` and `gdal`.
This commit adjusts the following dependencies declared in the
`GdalConan::requirements()` member function:

  * `libjpeg`
  * `sqlite3`

These now use an explicit version range, rather than an exact
version specification, in order to resolve conflicts with the
`osgearth` recipe as updated in #28101.

In addition, this commit also adds a new member function called
`GdalConan::validate_build()` which throws a
`ConanInvalidConfiguration` exception if the user attempts to build
GDAL with the MSVC using a C++ Standard which is greater than
C++17. This was done to account for build failures which were being
encountered in that particular case.

(NOTE: All of the changes discussed in this commit message apply to
the `GdalConan` class defined within
`recipes\gdal\pre_3.5.0\conanfile.py`. For versions of GDAL which
are at least as high as v3.5.0, a different recipe is used,
instead. This commit does not attempt to update that recipe.)