Fixed several issues reported by the Recipe Linter.
This commit addresses several issues which the Recipe Linter
diagnosed for the `osgearth` recipe. These mostly consist of unused
`import` statements, but it did warn about a missing `package_type`
attribute for `OsgearthConan`. (There was also an "error" for not
using `conan.tools.files.export_conandata_patches()`, although I
suspect that the previous use of `OsgearthConan.exports_sources`
achieved equivalent results.)

This commit also updates the `required_conan_version` value for the
`osgearth` recipe to `">=2.0"` based on code review feedback. It
still isn't quite clear if this is correct, though, because I have
not actually tested the recipe at that version.