openscenegraph: added some patches (#28121)
* Fixed several `openscenegraph` build failures.

This commit includes several additional patch files in order to
resolve build failures which occurred when attempting to build
`openscenegraph` v3.6.5 with the MSVC. An adjustment was also made
to the declared required `libjpeg` version; this was done to
resolve a dependency conflict for #28101.

I should mention that I have not done any significant analysis
regarding the implications of increasing the `libjpeg` version
requirement. This change resolves the aforementioned dependency
conflict, but it could also theoretically introduce new ones. It
may be better to replace this with a version range, instead.

* Modified the version specification for `libjpeg` in `openscenegraph`.

This commit updates the declared dependency for `libjpeg` within
`OpenSceneGraphConanFile::requirements()` to make use of a version
range, rather than an explicit version. This was done at the
request of @AbrilRBS in a code review comment.

* Add more version ranges, remove Conan 1 name usage

* Rename patches

---------

Co-authored-by: Abril Rincón Blanco <5364255+AbrilRBS@users.noreply.github.com>